Changed variable declaration on line 4 from int to float to produce decimal number as result.

+++ Calculator/Program.cs
@@ -1,7 +1,7 @@
 ﻿// See https://aka.ms/new-console-template for more information
 
 // Declare variables and then initialise to zero
-int num1 = 0; int num2 = 0;
+float num1 = 0; float num2 = 0;
 
 // Display title as the C# console calculator app.
 Console.WriteLine("Console Calculator in C#\r");
